fixed: delete user with no profile and wall

+++ app.py
@@ -1,21 +1,25 @@
 import os
+import secrets
+from datetime import date
 from functools import wraps
-from flask import Flask, render_template, redirect, url_for, flash, abort, send_from_directory
+
+from flask import Flask, render_template, redirect, url_for, flash, abort, send_from_directory, current_app
 from flask_bootstrap import Bootstrap
 from flask_ckeditor import CKEditor
-from datetime import date
-from flask_login import LoginManager, current_user, login_required
 from flask_gravatar import Gravatar
-import secrets
-from flask_session import Session
+from flask_login import LoginManager, current_user, login_required
+
 from db import User, BlogPost, db, Comment
-from forms import CommentForm, CreatePostForm, SearchForm
+from flask_session import Session
+from forms import CommentForm, CreatePostForm
 from functions import delete_file, img_to_uuid
 from log_reg import log_bp
-from others import others_bp
 from manage_user import user_bp
+from others import others_bp
+from modify_db import modify_db
+
+# from handle_error import handle_error_bp
 # from manage_post import post_bp
-from dotenv import load_dotenv
 
 app = Flask(__name__)
 app.app_context().push()
@@ -30,7 +34,9 @@ Bootstrap(app)
 
 # CONNECT TO DB
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///blog.db'
-app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
+# app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///blogs.db'
+app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = True
+
 UPLOAD_USER_IMG = "dynamic/profile_pic"
 app.config['UPLOAD_USER_IMG'] = UPLOAD_USER_IMG
 UPLOAD_BLOG_IMG = "dynamic/blog_img"
@@ -46,12 +52,12 @@ login_manager.init_app(app)
 
 # load_dotenv()
 # auth_users = (os.getenv('AUTH_USERS'))
-auth_users = [1, 2, 3]
+auth_users = [1, 2]
 
 
 @login_manager.user_loader
 def load_user(user_id):
-    return User.query.get_or_404(int(user_id))
+    return User.query.get_or_404(user_id)
 
 
 def admin_only(f):
@@ -67,6 +73,7 @@ def admin_only(f):
 app.register_blueprint(log_bp)
 app.register_blueprint(others_bp)
 app.register_blueprint(user_bp)
+# app.register_blueprint(handle_error_bp)
 # app.register_blueprint(post_bp)
 
 
@@ -84,6 +91,7 @@ year = date.today().year
 @app.route('/')
 def get_all_posts():
     posts = BlogPost.query.all()
+    # modify_db()
     return render_template("index.html", all_posts=posts, logged_in=current_user.is_authenticated, year=year,
                            auth_users=auth_users)
 
@@ -113,6 +121,7 @@ def uploaded_file(filename):
 def show_post(post_id):
     if current_user.is_authenticated:
         requested_post = BlogPost.query.get_or_404(post_id)
+        # all_comments = Comment.query.filter_by(post_id=requested_post.id).all()
         form = CommentForm()
         if form.validate_on_submit():
             if not current_user.id:
@@ -124,12 +133,13 @@ def show_post(post_id):
             )
             db.session.add(new_comment)
             db.session.commit()
-            return redirect(url_for('get_all_posts'))
-
-        img_url = url_for('uploaded_file', filename=requested_post.img_url)
+            return redirect(url_for('show_post', post_id=post_id))
+        img_url = None
+        if requested_post.img_url:
+            img_url = url_for('uploaded_file', filename=requested_post.img_url)
 
         return render_template("post.html", post=requested_post, form=form, current_user=current_user, img_url=img_url,
-                               year=year)
+                               year=year, )
     flash("You're not logged-in. Kindly Log-in")
     return redirect(url_for('get_all_posts'))
 
@@ -139,19 +149,34 @@ def show_post(post_id):
 def add_new_post():
     form = CreatePostForm()
     if form.validate_on_submit():
-        img = form.img_url
-        img_id = img_to_uuid(img)
-        img.data.save(os.path.join(app.config['UPLOAD_BLOG_IMG'], img_id))
-        new_post = BlogPost(
-            title=form.title.data,
-            subtitle=form.subtitle.data,
-            body=form.body.data,
-            img_url=img_id,
-            author=current_user,
-            date=date.today().strftime("%B %d, %Y")
-        )
-        db.session.add(new_post)
-        db.session.commit()
+        print("valid")
+        if form.img_url.data:
+            img = form.img_url
+            img_id = img_to_uuid(img)
+            img.data.save(os.path.join(app.config['UPLOAD_BLOG_IMG'], img_id))
+            new_post = BlogPost(
+                title=form.title.data,
+                subtitle=form.subtitle.data,
+                body=form.body.data,
+                img_url=img_id,
+                author=current_user,
+                date=date.today().strftime("%B %d, %Y")
+            )
+            db.session.add(new_post)
+            db.session.commit()
+            print("form")
+        else:
+            print("else")
+            new_post = BlogPost(
+                title=form.title.data,
+                subtitle=form.subtitle.data,
+                body=form.body.data,
+                img_url=None,
+                author=current_user,
+                date=date.today().strftime("%B %d, %Y")
+            )
+            db.session.add(new_post)
+            db.session.commit()
         return redirect(url_for("get_all_posts"))
     return render_template("make-post.html", form=form, year=year)
 
@@ -160,6 +185,8 @@ def add_new_post():
 @login_required
 def edit_post(post_id):
     post = BlogPost.query.get_or_404(post_id)
+    file_path = os.path.join(UPLOAD_BLOG_IMG, post.img_url) if post.img_url else None
+
     edit_form = CreatePostForm(
         title=post.title,
         subtitle=post.subtitle,
@@ -167,19 +194,23 @@ def edit_post(post_id):
         author=post.author,
         body=post.body
     )
-    file_path = os.path.join(UPLOAD_BLOG_IMG, post.img_url)
     if edit_form.validate_on_submit():
-        img = edit_form.img_url
-        img_id = img_to_uuid(img)
-        img.data.save(os.path.join(app.config['UPLOAD_BLOG_IMG'], img_id))
 
         post.title = edit_form.title.data
         post.subtitle = edit_form.subtitle.data
-        post.img_url = img_id
         post.body = edit_form.body.data
+
+        if edit_form.img_url:
+            img = edit_form.img_url
+            img_id = img_to_uuid(img)
+            img.data.save(os.path.join(app.config['UPLOAD_BLOG_IMG'], img_id))
+            post.img_url = img_id
+
+            if file_path:
+                delete_file(file_path)
+
         db.session.commit()
 
-        delete_file(file_path)
         return redirect(url_for("show_post", post_id=post_id))
 
     return render_template("make-post.html", form=edit_form, year=year)
@@ -202,16 +233,41 @@ def delete_post(post_id):
         return redirect(url_for('get_all_posts'))
 
 
+@app.route("/delete_comment/<int:cmt_id>", methods=["GET", "POST"])
+@login_required
+def delete_cmt(cmt_id):
+    cmt_to_delete = Comment.query.get_or_404(cmt_id)
+    try:
+        Comment.query.filter_by(post_id=cmt_id).delete()
+        db.session.delete(cmt_to_delete)
+        db.session.commit()
+        flash("Comment was deleted.")
+        return redirect(url_for('show_post', post_id=cmt_to_delete.post_id))
+    except:
+        flash("Whoops!! There was a problem deleting that comment.")
+        return redirect(url_for('show_post', post_id=cmt_to_delete.post_id))
+
+
 @app.errorhandler(404)
 def page_not_found(e):
     return render_template("404.html"), 404
 
 
-# Internal Server Error
 @app.errorhandler(500)
 def page_not_found(e):
-    return render_template("500.html"), 500
+    return render_template("500.html"), 500 @ app.errorhandler(404)
+
+
+@app.errorhandler(401)
+def page_not_found(e):
+    return render_template("401.html"), 401
+
+
+@app.errorhandler(403)
+def page_not_found(e):
+    return render_template("403.html"), 403
 
 
 if __name__ == "__main__":
     app.run(debug=True)
+# serve(app, host='0.0.0.0', port=80, threads=1)
